Update Reflex.py
I edited the code so that the up button, rather than the down button, skips from the screen that displays the name of the shape to the gameplay screen that shows the shapes. This is primarily because the thumby library does not seem to have a function to detect a button release (just whether a specific button is currently pressed down or has most recently been pressed down).