Update Reflex.py
I added an anti-cheating mechanism. Previously, if you were holding down the A button when the correct shape appeared on the screen, you would immediately advance to the end screen with a very short reaction time. Now, when the correct shape appears, the game will not advance to the end screen until the program ensures that the A button was not pressed down when the shape first appeared. The game will only proceed to the end if the player releases the button (or has already released the button) and presses it for the shape.