Update Reflex.py
I debugged an issue in the game that appeared after I made the beginStopwatch() function. Every function call was not modifying the global copy of the begin variable, so the statement "react_time = end - begin" would use the original value of begin (set to 0 at the beginning of the game). Now the statement "global begin" ensures that beginStopwatch() is modifying the begin variable for the whole program.